+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added cmdlet `Compare-DscParameterState` - Could be used in
+  Get-TargetResource function or Get() method in Class based Resources.
+  It is based on the code of Test-DscParameterState function to get compliance between current and
+  desired state of resources.
+
+### Changed
+
+- Cmdlet Test-DscResourceState is now a wrapper of Compare-DscParameterState. Possible breaking change.
+
 ## [0.9.3] - 2020-07-25
 
 ## Fixed