+++ CHANGELOG.md
@@ -11,8 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added cmdlet `Compare-DscParameterState` - Could be used in
   Get-TargetResource function or Get() method in Class based Resources.
-  It is based on the code of Test-DscParameterState function to get compliance between current and
-  desired state of resources.
+  It is based on the code of Test-DscParameterState function to get compliance
+  between current and desired state of resources.
+- Added pester test to test the pscredential object with `Compare-DscParameterState`
 
 ### Changed